refactor(azure): extract shared hosted-DB scanner into azure_db_common
