+++ pipeline/ingest/azure_sql.py
@@ -14,104 +14,17 @@ from collections.abc import Iterable
 from pathlib import Path
 from typing import Any
 
-from normalize.enums import apply_kind_defaults
-from normalize.terms import terms_hash
-
-from ._duckdb import dumps, open_conn
-from .azure_common import load_region_normalizer, parse_unit_of_measure
-
-_PROVIDER = "azure"
-_SERVICE = "sql"
-_KIND = "db.relational"
-
-_DEPLOYMENT_HINTS: tuple[tuple[str, str], ...] = (
-    ("Managed Instance", "managed-instance"),
-    ("Elastic Pool",     "elastic-pool"),
-    ("Single Database",  "single-az"),
-)
-
-_SQL = """
-WITH items AS (
-  SELECT UNNEST(Items, recursive := true)
-  FROM read_json_auto('{path}', maximum_object_size=33554432)
-)
-SELECT
-  CAST(meterId AS VARCHAR) AS sku_id,
-  skuName       AS resource_name,
-  armRegionName AS region,
-  productName   AS product_name,
-  retailPrice   AS price,
-  unitOfMeasure AS uom,
-  currencyCode  AS currency,
-  type          AS row_type,
-  serviceName   AS service_name
-FROM items
-"""
-
-
-def _classify_deployment(product_name: str) -> str | None:
-    for hint, value in _DEPLOYMENT_HINTS:
-        if hint in product_name:
-            return value
-    return None
+from ._duckdb import dumps
+from .azure_db_common import ingest_hosted_db
 
 
 def ingest(*, prices_path: Path) -> Iterable[dict[str, Any]]:
-    normalizer = load_region_normalizer()
-    con = open_conn()
-    path_literal = str(prices_path).replace("'", "''")
-    sql = _SQL.replace("{path}", path_literal)
-    seen: set[str] = set()
-    for (
-        sku_id, sku_name, region, product, price, uom, currency, row_type, service_name,
-    ) in con.execute(sql).fetchall():
-        if service_name != "SQL Database":
-            continue
-        if row_type != "Consumption":
-            continue
-        if currency != "USD":
-            continue
-        deployment = _classify_deployment(product)
-        if deployment is None:
-            continue
-        region_normalized = normalizer.try_normalize(_PROVIDER, region)
-        if region_normalized is None:
-            continue
-        try:
-            divisor, unit = parse_unit_of_measure(uom)
-        except ValueError:
-            continue  # skip non-hourly meters (monthly, per-request, etc.)
-        if sku_id in seen:
-            continue
-        seen.add(sku_id)
-        terms = apply_kind_defaults(_KIND, {
-            "commitment": "on_demand",
-            "tenancy": "azure-sql",
-            "os": deployment,
-            "support_tier": "",
-            "upfront": "",
-            "payment_option": "",
-        })
-        yield {
-            "sku_id": sku_id,
-            "provider": _PROVIDER,
-            "service": _SERVICE,
-            "kind": _KIND,
-            "resource_name": sku_name,
-            "region": region,
-            "region_normalized": region_normalized,
-            "terms_hash": terms_hash(terms),
-            "resource_attrs": {
-                "extra": {
-                    "product_name": product,
-                    "deployment_option": deployment,
-                },
-            },
-            "terms": terms,
-            "prices": [
-                {"dimension": "compute", "tier": "", "amount": price / divisor, "unit": unit},
-            ],
-        }
+    yield from ingest_hosted_db(
+        prices_path=prices_path,
+        service_name="SQL Database",
+        tenancy_slug="azure-sql",
+        service="sql",
+    )
 
 
 def main() -> int:
